更新: [class Log]新增Debug选项: 显示Debug项目模式下的Debug信息;修改后台日志写入线程的退出模式: 主线程发出停止消息后, 等待后台线程退出并输出提示信息. [namespace Time]修改秒/毫秒/微秒/纳秒的字符转换为英文s/ms/us/ns. [class FileSystem]新增std::filesystem的异常处理和异常信息的输出.

+++ src/WindowsHost.cpp
@@ -277,7 +277,7 @@ Typical_Tool::WindowsSystem::ShellMessage Typical_Tool::WindowsSystem::WindowShe
 	}
 	lgc(_T("ExecuteAnalyze: 窗口显示 ") + 窗口显示, wr);
 
-	ShellMessage temp(操作名, (int)ShellExecuteW(NULL, stow(Shell操作).c_str(), stow(Shell文件).c_str(), stow(Shell参数).c_str(), NULL, ShowWindow));
+	ShellMessage temp(操作名, (long long)ShellExecuteW(NULL, stow(Shell操作).c_str(), stow(Shell文件).c_str(), stow(Shell参数).c_str(), NULL, ShowWindow));
 	return temp;
 }
